Create AntiAttacks-PP.py
This is a Personal Protection for Online threats made against you. IT shows much like TCP dumps but further providing accurate logs in a txt file, cmd prompt or ssh. 

It's only a watching activities incoming that are threats for you to use.

There is only one way to help others, that's to place something usable out there! Enjoy.