Update AntiAttacks-PP.py
Updated to obtain the local ip of machine running on instead of watching 127 or 0.0.0.0 which is 127 by default. 

Added sniff package also

This is a Personal Protection for Online threats made against you. IT shows much like TCP dumps but further providing accurate logs in a txt file, cmd prompt or ssh. 

It's only a watching activities incoming that are threats for you to use.

There is only one way to help others, that's to place something usable out there! Enjoy.